+++ eva/src/components/dashPanelBoard.vue
@@ -2011,8 +2011,7 @@ export default {
           events.forEach((item) => {
             originItem = item;
             if (item !== '') {
-              // reg = new RegExp(/^[\s+]?[\w]+\(/, 'g');
-              reg = /^[\s+]?[\w]+\(/g;
+              reg = /^\s*\w+\(/g;
               this.$set(
                 this.event,
                 'event',